#5785 Extract method for calculating compaction.
