+++ ApplicationCode/GeoMech/GeoMechDataModel/RigFemPartResultsCollection.cpp
@@ -33,6 +33,7 @@
 #include "RigFemNativeStatCalc.h"
 #include "RigFemPartCollection.h"
 #include "RigFemPartGrid.h"
+#include "RigFemPartResultCalculatorCompaction.h"
 #include "RigFemPartResultCalculatorDSM.h"
 #include "RigFemPartResultCalculatorED.h"
 #include "RigFemPartResultCalculatorEV.h"
@@ -55,7 +56,6 @@
 #include "RigFemScalarResultFrames.h"
 #include "RigFormationNames.h"
 #include "RigHexGradientTools.h"
-#include "RigHexIntersectionTools.h"
 #include "RigStatisticsDataCache.h"
 #include "RigWbsParameter.h"
 
@@ -67,7 +67,6 @@
 #include "cafProgressInfo.h"
 #include "cafTensor3.h"
 
-#include "cvfBoundingBox.h"
 #include "cvfMath.h"
 
 #include <QString>
@@ -79,23 +78,6 @@
 ///
 //--------------------------------------------------------------------------------------------------
 const std::string RigFemPartResultsCollection::FIELD_NAME_COMPACTION = "COMPACTION";
-
-//--------------------------------------------------------------------------------------------------
-/// Internal definitions
-//--------------------------------------------------------------------------------------------------
-class RefElement
-{
-public:
-    size_t              elementIdx;
-    float               intersectionPointToCurrentNodeDistance;
-    cvf::Vec3f          intersectionPoint;
-    std::vector<size_t> elementFaceNodeIdxs;
-};
-
-static std::vector<cvf::Vec3d> coordsFromNodeIndices( const RigFemPart& part, const std::vector<size_t>& nodeIdxs );
-static std::vector<size_t>     nodesForElement( const RigFemPart& part, size_t elementIdx );
-static float                   horizontalDistance( const cvf::Vec3f& p1, const cvf::Vec3f& p2 );
-static void findReferenceElementForNode( const RigFemPart& part, size_t nodeIdx, size_t kRefLayer, RefElement* refElement );
 
 //--------------------------------------------------------------------------------------------------
 ///
@@ -131,6 +113,8 @@ RigFemPartResultsCollection::RigFemPartResultsCollection( RifGeoMechReaderInterf
         std::shared_ptr<RigFemPartResultCalculator>( new RigFemPartResultCalculatorSurfaceAngles( *this ) ) );
     m_resultCalculators.push_back(
         std::shared_ptr<RigFemPartResultCalculator>( new RigFemPartResultCalculatorSurfaceAlignedStress( *this ) ) );
+    m_resultCalculators.push_back(
+        std::shared_ptr<RigFemPartResultCalculator>( new RigFemPartResultCalculatorCompaction( *this ) ) );
     m_resultCalculators.push_back(
         std::shared_ptr<RigFemPartResultCalculator>( new RigFemPartResultCalculatorSFI( *this ) ) );
     m_resultCalculators.push_back(
@@ -1078,84 +1062,6 @@ RigFemScalarResultFrames* RigFemPartResultsCollection::calculateNodalGradients( 
 //--------------------------------------------------------------------------------------------------
 ///
 //--------------------------------------------------------------------------------------------------
-RigFemScalarResultFrames* RigFemPartResultsCollection::calculateCompactionValues( int                        partIndex,
-                                                                                  const RigFemResultAddress& resVarAddr )
-{
-    CVF_ASSERT( resVarAddr.fieldName == FIELD_NAME_COMPACTION );
-
-    caf::ProgressInfo frameCountProgress( this->frameCount() + 1, "" );
-    frameCountProgress.setProgressDescription( "Calculating " + QString::fromStdString( resVarAddr.fieldName ) );
-
-    RigFemScalarResultFrames* u3Frames =
-        this->findOrLoadScalarResult( partIndex, RigFemResultAddress( resVarAddr.resultPosType, "U", "U3" ) );
-    frameCountProgress.incrementProgress();
-
-    RigFemScalarResultFrames* compactionFrames = m_femPartResults[partIndex]->createScalarResult( resVarAddr );
-
-    const RigFemPart* part = m_femParts->part( partIndex );
-    part->ensureIntersectionSearchTreeIsBuilt();
-
-    for ( int t = 0; t < u3Frames->frameCount(); t++ )
-    {
-        std::vector<float>& compactionFrame = compactionFrames->frameData( t );
-        size_t              nodeCount       = part->nodes().nodeIds.size();
-
-        frameCountProgress.incrementProgress();
-
-        compactionFrame.resize( nodeCount );
-
-        {
-            // Make sure the AABB-tree is created before using OpenMP
-            cvf::BoundingBox    bb;
-            std::vector<size_t> refElementCandidates;
-
-            part->findIntersectingCells( bb, &refElementCandidates );
-
-            // Also make sure the struct grid is created, as this is required before using OpenMP
-            part->getOrCreateStructGrid();
-        }
-
-#pragma omp parallel for
-        for ( long n = 0; n < static_cast<long>( nodeCount ); n++ )
-        {
-            RefElement refElement;
-            findReferenceElementForNode( *part, n, resVarAddr.refKLayerIndex, &refElement );
-
-            if ( refElement.elementIdx != cvf::UNDEFINED_SIZE_T )
-            {
-                float  shortestDist      = std::numeric_limits<float>::infinity();
-                size_t closestRefNodeIdx = cvf::UNDEFINED_SIZE_T;
-
-                for ( size_t nodeIdx : refElement.elementFaceNodeIdxs )
-                {
-                    float dist = horizontalDistance( refElement.intersectionPoint, part->nodes().coordinates[nodeIdx] );
-                    if ( dist < shortestDist )
-                    {
-                        shortestDist      = dist;
-                        closestRefNodeIdx = nodeIdx;
-                    }
-                }
-
-                cvf::Vec3f currentNodeCoord = part->nodes().coordinates[n];
-                if ( currentNodeCoord.z() >= refElement.intersectionPoint.z() )
-                    compactionFrame[n] = -( u3Frames->frameData( t )[n] - u3Frames->frameData( t )[closestRefNodeIdx] );
-                else
-                    compactionFrame[n] = -( u3Frames->frameData( t )[closestRefNodeIdx] - u3Frames->frameData( t )[n] );
-            }
-            else
-            {
-                compactionFrame[n] = HUGE_VAL;
-            }
-        }
-    }
-
-    RigFemScalarResultFrames* requestedPrincipal = this->findOrLoadScalarResult( partIndex, resVarAddr );
-    return requestedPrincipal;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
 RigFemScalarResultFrames* RigFemPartResultsCollection::calculateNE( int partIndex, const RigFemResultAddress& resVarAddr )
 {
     caf::ProgressInfo frameCountProgress( this->frameCount() * 2, "" );
@@ -1262,11 +1168,6 @@ RigFemScalarResultFrames* RigFemPartResultsCollection::calculateDerivedResult( i
     for ( auto calculator : m_resultCalculators )
     {
         if ( calculator->isMatching( resVarAddr ) ) return calculator->calculate( partIndex, resVarAddr );
-    }
-
-    if ( resVarAddr.fieldName == FIELD_NAME_COMPACTION )
-    {
-        return calculateCompactionValues( partIndex, resVarAddr );
     }
 
     if ( resVarAddr.fieldName == "ST" || resVarAddr.fieldName == "SE" )
@@ -1934,105 +1835,6 @@ int RigFemPartResultsCollection::partCount() const
 const RigFemPartCollection* RigFemPartResultsCollection::parts() const
 {
     return m_femParts.p();
-}
-
-//--------------------------------------------------------------------------------------------------
-/// Internal functions
-//--------------------------------------------------------------------------------------------------
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<cvf::Vec3d> coordsFromNodeIndices( const RigFemPart& part, const std::vector<size_t>& nodeIdxs )
-{
-    std::vector<cvf::Vec3d> out;
-    for ( const auto& nodeIdx : nodeIdxs )
-        out.push_back( cvf::Vec3d( part.nodes().coordinates[nodeIdx] ) );
-    return out;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<size_t> nodesForElement( const RigFemPart& part, size_t elementIdx )
-{
-    std::vector<size_t> nodeIdxs;
-    const int*          nodeConn = part.connectivities( elementIdx );
-    for ( int n = 0; n < 8; n++ )
-        nodeIdxs.push_back( nodeConn[n] );
-    return nodeIdxs;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-float horizontalDistance( const cvf::Vec3f& p1, const cvf::Vec3f& p2 )
-{
-    cvf::Vec3f p1_ = p1;
-    cvf::Vec3f p2_ = p2;
-    p1_.z() = p2_.z() = 0;
-    return p1_.pointDistance( p2_ );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void findReferenceElementForNode( const RigFemPart& part, size_t nodeIdx, size_t kRefLayer, RefElement* refElement )
-{
-    static const double zMin = -1e6, zMax = 1e6;
-
-    cvf::BoundingBox bb;
-    cvf::Vec3f       currentNodeCoord = part.nodes().coordinates[nodeIdx];
-    cvf::Vec3f       p1               = cvf::Vec3f( currentNodeCoord.x(), currentNodeCoord.y(), zMin );
-    cvf::Vec3f       p2               = cvf::Vec3f( currentNodeCoord.x(), currentNodeCoord.y(), zMax );
-    bb.add( p1 );
-    bb.add( p2 );
-
-    std::vector<size_t> refElementCandidates;
-    part.findIntersectingCells( bb, &refElementCandidates );
-
-    const RigFemPartGrid* grid = part.getOrCreateStructGrid();
-
-    refElement->elementIdx                             = cvf::UNDEFINED_SIZE_T;
-    refElement->intersectionPointToCurrentNodeDistance = std::numeric_limits<float>::infinity();
-    size_t i, j, k;
-    for ( const size_t elemIdx : refElementCandidates )
-    {
-        bool validIndex = grid->ijkFromCellIndex( elemIdx, &i, &j, &k );
-        if ( validIndex && k == kRefLayer )
-        {
-            const std::vector<size_t> nodeIndices = nodesForElement( part, elemIdx );
-            CVF_ASSERT( nodeIndices.size() == 8 );
-
-            std::vector<HexIntersectionInfo> intersections;
-            RigHexIntersectionTools::lineHexCellIntersection( cvf::Vec3d( p1 ),
-                                                              cvf::Vec3d( p2 ),
-                                                              coordsFromNodeIndices( part, nodeIndices ).data(),
-                                                              elemIdx,
-                                                              &intersections );
-
-            for ( const auto& intersection : intersections )
-            {
-                cvf::Vec3f intersectionPoint = cvf::Vec3f( intersection.m_intersectionPoint );
-
-                float nodeToIntersectionDistance = currentNodeCoord.pointDistance( intersectionPoint );
-                if ( nodeToIntersectionDistance < refElement->intersectionPointToCurrentNodeDistance )
-                {
-                    cvf::ubyte faceNodes[4];
-                    grid->cellFaceVertexIndices( intersection.m_face, faceNodes );
-                    std::vector<size_t> topFaceCoords( {nodeIndices[faceNodes[0]],
-                                                        nodeIndices[faceNodes[1]],
-                                                        nodeIndices[faceNodes[2]],
-                                                        nodeIndices[faceNodes[3]]} );
-
-                    refElement->elementIdx                             = elemIdx;
-                    refElement->intersectionPointToCurrentNodeDistance = nodeToIntersectionDistance;
-                    refElement->intersectionPoint                      = intersectionPoint;
-                    refElement->elementFaceNodeIdxs                    = topFaceCoords;
-                }
-            }
-        }
-    }
 }
 
 //--------------------------------------------------------------------------------------------------
